# lib/api-zod/src/generated/types/bookOddsEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Betting Predictions API — powered by Rover
 * OpenAPI spec version: 0.1.0
 */

export interface BookOddsEntry {
  id: number;
  gameId: number;
  bookName: string;
  homeMoneyline: number;
  awayMoneyline: number;
  homeSpread: number;
  awaySpread: number;
  spreadJuice: number;
  overUnder: number;
  overJuice: number;
  underJuice: number;
  lastUpdated: string;
}