# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Betting Predictions API — powered by Rover
 * OpenAPI spec version: 0.1.0
 */
import type { GameOdds } from "./gameOdds";
import type { GameStatus } from "./gameStatus";

export interface Game {
  id: number;
  sportId: number;
  sportName: string;
  homeTeam: string;
  awayTeam: string;
  scheduledAt: string;
  status: GameStatus;
  league: string;
  /** @nullable */
  homeScore?: number | null;
  /** @nullable */
  awayScore?: number | null;
  /** @nullable */
  predictionId?: number | null;
  odds?: GameOdds;
}