# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports Betting Predictions API — powered by Rover
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalPicks: number;
  wins: number;
  losses: number;
  pushes: number;
  winRate: number;
  totalGames: number;
  upcomingGames: number;
  activePredictions: number;
  unitPnl: number;
  roi: number;
  edgeHits: number;
}